requirements.txt
nginx
복사편집
streamlit
pytesseract
Pillow
pandas